Add initial project structure and configuration files
- Create CMakeLists.txt for backend build configuration
- Add .gitignore to exclude build artifacts and environment files
- Include LICENSE and README files for project documentation
- Set up dataset pipeline and API documentation
- Implement environment variables in .env.example
- Add PostCSS configuration for frontend styling
- Create placeholder files for datasets, tests, and screenshots

+++ backend/src/algorithms.cpp
@@ -142,7 +142,7 @@ PathResult ShortestPathService::aStar(int source, int target) const{
 PathResult ShortestPathService::bellmanFord(int source, int target) const{
     const auto started = Clock::now();
     const int n = graph_.size();
-    std::vector<int> dist(n, kInf);
+    std::vector<double> dist(n, kInf);
     std::vector<int> parent(n, -1);
     long long states = 0;
